+++ R/heddle.R
@@ -7,8 +7,10 @@
 #'
 #' @param data Input dataframe to pull replacement values from. Accepts either
 #' vector or dataframe inputs.
-#' @param pattern The base pattern, which will be replicated for each value in
-#' the vector with keywords replaced by data.
+#' @param pattern The base pattern, as either an atomic vector (which will be
+#' recycled for every value in your data) or a vector of the same length as
+#' your data (which will be applied element-wise to your data, so that
+#' \code{data[[1]]} will replace \code{pattern[[1]]}).
 #' @param ... If \code{data} is a vector, a string representing the string to
 #' replace with data values in each pattern. If \code{data} is a dataframe, a
 #' set of name = variable pairs, with the name matching the keyword to be
@@ -45,6 +47,10 @@ heddle <- function(data, pattern, ..., strip.whitespace = FALSE) {
 #' @export
 heddle.default <- function(data, pattern, ..., strip.whitespace = FALSE) {
   dots <- list(...)
+  if (length(pattern) != 1 & length(pattern) != length(data)) {
+    stop("Argument pattern must be an atomic vector or have the same number of 
+         elements as your data.")
+  }
   if (length(dots) == 0) stop("argument '...' is missing, with no default")
   if (length(dots) > 1) {
     dots <- paste0(unlist(dots), collapse = "|")
@@ -57,12 +63,19 @@ heddle.default <- function(data, pattern, ..., strip.whitespace = FALSE) {
   } else if (strip.whitespace) {
     data <- gsub("[[:space:]]", "", data)
   }
-  pat <- pattern
-  vapply(
-    data,
-    function(x) gsub(dots, x, pat),
-    FUN.VALUE = character(1)
-  )
+  if (length(pattern) == 1) {
+    vapply(
+      data,
+      function(x) gsub(dots, x, pattern),
+      FUN.VALUE = character(1)
+    )
+  } else {
+    mapply(function(x, y) gsub(dots, x, y),
+      x = data,
+      y = pattern,
+      SIMPLIFY = FALSE
+    )
+  }
 }
 
 #' @export