+++ R/heddle.R
@@ -1,7 +1,7 @@
 #' Transform pattern objects into template pieces
 #'
-#' This function replicates pattern objects, replacing keywords in each
-#' iteration with values from the provided data. This allows efficiently
+#' This function replicates pattern objects, replacing placeholder keywords in
+#' each iteration with values from the provided data. This allows efficiently
 #' creating R Markdown documents with many repeating pieces which may shift
 #' alongside the underlying data.
 #'