+++ R/import_pattern.R
@@ -13,7 +13,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' export_template("my sample pattern", out.Rmd)
+#' export_template("my sample pattern", "out.Rmd")
 #' import_template("out.Rmd")
 #' }
 #'