+++ tests/testthat/test-import_draft.R
@@ -1,16 +1,20 @@
-test_that("assemble_draft fails when it should", {
-  expect_error(assemble_draft())
-  expect_error(assemble_draft(iris))
-  expect_warning(assemble_draft(c("pattern" = "../rmd/empty_file.Rmd")))
-  draft <- assemble_draft(
+test_that("import_draft fails when it should", {
+  expect_error(import_draft())
+  expect_error(import_draft(iris))
+  expect_warning(import_draft(c("pattern" = "../rmd/empty_file.Rmd")))
+  draft <- import_draft(
     "one" = "../rmd/sample_pattern.Rmd",
     "two" = "../rmd/sample_pattern.Rmd"
   )
   expect_error(names(draft)[[3]])
+  expect_warning(assemble_draft(
+    "one" = "../rmd/sample_pattern.Rmd",
+    "two" = "../rmd/sample_pattern.Rmd"
+  ))
 })
 
-test_that("assemble_draft imports objects as expected", {
-  draft <- assemble_draft(
+test_that("import_draft imports objects as expected", {
+  draft <- import_draft(
     "one" = "../rmd/sample_pattern.Rmd",
     "two" = "../rmd/sample_pattern.Rmd"
   )